Fix counselors 403'ing on their own leads with a strict-type comparison
Root cause of the /matchmaker/clients/1 403: authorizeClient() compared
$lead->assigned_to === auth()->id() with PHP's strict ===. The lead's
own My Clients list uses a plain SQL WHERE for the same check, which
MySQL coerces (assigned_to = '253' matches 253) — but the strict PHP
comparison doesn't, so an account could see its own lead correctly in
the list yet still get 403'd opening it directly. Cast both sides to
(int) everywhere a lead/profile ownership check compares against
auth()->id() — web, API, and the two assigned-to dropdowns' "selected"
state, which had the identical bug for a different symptom (wrong
counselor shown as selected).

+++ app/Http/Controllers/Api/V1/Matchmaker/ClientController.php
@@ -47,7 +47,10 @@ class ClientController extends Controller
             return;
         }
 
-        abort_unless($lead->assigned_to === auth()->id(), 403, 'This client is assigned to another Nikah Counselor, so it is hidden from your account for privacy. If this client should be yours, ask your admin to reassign it to you.');
+        // (int) cast on both sides — see the web ClientController's
+        // authorizeClient() for why a strict === here can wrongly 403 an
+        // account whose own list still correctly shows the same lead.
+        abort_unless((int) $lead->assigned_to === (int) auth()->id(), 403, 'This client is assigned to another Nikah Counselor, so it is hidden from your account for privacy. If this client should be yours, ask your admin to reassign it to you.');
     }
 
     public function index(Request $request): JsonResponse
